third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFETurbulenceElement-svgdom-type-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'type' property of the SVGFETurbulenceElment object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS turbulence.type.baseVal is SVGFETurbulenceElement.SVG_TURBULENCE_TYPE_FRACTALNOISE
PASS turbulence.type.baseVal is SVGFETurbulenceElement.SVG_TURBULENCE_TYPE_TURBULENCE
PASS successfullyParsed is true

TEST COMPLETE
